[TEST] Testing and bug fixes have been performed. Some documentation has been added.
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/msnovogen@11672 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/DENOVO/MSNOVOGEN/DefaultMutater.C
@@ -40,7 +40,7 @@ namespace OpenMS
 	DefaultMutater::DefaultMutater(double precursorMass, double precursorMassTolerance, std::vector<const Residue*> aaList) :
 		Mutater(precursorMass, precursorMassTolerance, aaList), dm_(precursorMass, precursorMassTolerance, aaList)
 	{
-		dm_.seed(time(0));
+		dm_.seed((unsigned int)time(0));
 	}
 
 	DefaultMutater::~DefaultMutater()
